refactor: split App.jsx into hooks/components and fix backend URL for external access
Backend WS/REST URLs now derived from window.location.hostname at runtime instead of
hardcoded build-time IP, fixing graph panel for users accessing via NAT port forwarding.
