+++ chat_front/Dockerfile
@@ -3,6 +3,10 @@ WORKDIR /app
 COPY package.json package-lock.json* ./
 RUN npm install
 COPY . .
+ARG VITE_WORKFLOW_WS_URL
+ARG VITE_WORKFLOW_GRAPH_URL
+ENV VITE_WORKFLOW_WS_URL=$VITE_WORKFLOW_WS_URL
+ENV VITE_WORKFLOW_GRAPH_URL=$VITE_WORKFLOW_GRAPH_URL
 RUN npm run build
 EXPOSE 10000
 CMD ["npm", "run", "preview", "--", "--host", "0.0.0.0", "--port", "10000"]
