+++ chat_front/vite.config.js
@@ -2,5 +2,8 @@ import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
 export default defineConfig({
-  plugins: [react()]
+  plugins: [react()],
+  preview: {
+    allowedHosts: ['oddbie.iptime.org'],
+  },
 })
